Rename `builder_nse.rs` to `effects.rs`
